20.把drawable 加载到 seekbar 看下效果。应该很丑.-

+++ app/src/main/java/com/utsoft/jan/myqqview/douyin/common/player/VideoImpl.java
@@ -25,6 +25,6 @@ public interface VideoImpl {
     int initTrack(MediaExtractor extractor);
 
     //编译出图片
-    Bitmap getBitmapBySec(MediaExtractor extractor, MediaFormat format, MediaCodec codec, int[] sec);
+    Bitmap getBitmapBySec(MediaExtractor extractor, MediaFormat format, MediaCodec codec, long[] sec);
 
 }
